% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbital_speed.R
\name{calc_obtital_speeds}
\alias{calc_obtital_speeds}
\title{Compute the orbital speed at a given depth from the wave elevation 1D spectra}
\usage{
calc_obtital_speeds(S, freq, z = 0, depth = Inf, spec = FALSE)
}
\arguments{
\item{S}{1D spectral data: TxM matrix}

\item{freq}{the M frequencies}

\item{z}{distance above the flood at which we want the orbital speed}

\item{depth}{depth time series (vector length T)}

\item{spec}{TRUE if the spectral speed are needed. Oterwise, returns the RMSE (default)}
}
\value{
depending on spec, a lsit or a data.frame
}
\description{
Compute the orbital speed at a given depth from the wave elevation 1D spectra
}
