% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastTrapz}
\alias{fastTrapz}
\title{Fast implementation of pracma::trapz from the Armadillo C++ library}
\usage{
fastTrapz(X, Y, dim)
}
\arguments{
\item{dim}{an integer giving the subscripts which the function will be applied over. 1 indicates rows, 2 indicates columns}

\item{x}{x-coordinates of points on the x-axis (vector)}

\item{y}{y-coordinates of function values (matrix)}
}
\description{
Compute the area of a multivariate function with (matrix) values Y at the points x.
}
