% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_data_download.R
\name{get_1Dspectrum}
\alias{get_1Dspectrum}
\title{Download the 1D spectrum data from IFREMER ftp}
\usage{
get_1Dspectrum(point, start = "1994-01-01", end = "1994-02-28")
}
\arguments{
\item{point}{the location name (string) requested. Alternatively, the node number. The consistency is checked internally.}

\item{start}{the starting date (as a string). The consistency is checked internally.}

\item{end}{the ending date as a string}
}
\value{
A list with the data read from the downloaded netCDF file.
}
\description{
Download the 1D spectrum data from IFREMER ftp
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
   spec1D = get_1Dspectrum("SEMREVO",start="1994-01-01",end="1994-02-28")
   r <- as.POSIXct(round(range(spec1D$forcings$time), "month"))
   image(spec1D$forcings$time,spec1D$freq,t(spec1D$ef),
          xaxt = "n",xlab='Time',
          ylab = "Frequency (Hz)")
   axis.POSIXct(1,spec1D$forcings$time,at=seq(r[1], r[2], by = "week"),
   format = "\%Y-\%m-\%d",las=2)
\dontshow{\}) # examplesIf}
}
