% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_data_download.R
\name{get_2d_spectrum_raw}
\alias{get_2d_spectrum_raw}
\title{Download the 2D spectrum data from IFREMER ftp for a single data.}
\usage{
get_2d_spectrum_raw(point, year, month)
}
\arguments{
\item{point}{the location name (string) requested.}

\item{year}{the year (as a string) requested.}

\item{month}{the month number, as a string,}
}
\value{
a list with the sea-state spectrum and forcings
}
\description{
No consistency checks are made, this function should not be called directly by the user
}
\keyword{internal}
