% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_data_download.R
\name{get_2Dspectrum}
\alias{get_2Dspectrum}
\title{Download the 2D spectrum data from IFREMER ftp}
\usage{
get_2Dspectrum(point, start = "1994-01-01", end = "1994-02-28")
}
\arguments{
\item{point}{the location name (string) requested. Alternatively, the node number. The consistency is checked internally.}

\item{start}{the starting date (as a string). The consistency is checked internally.}

\item{end}{the ending date as a string}
}
\value{
A list with the data read from the downloaded netCDF file.
}
\description{
Download the 2D spectrum data from IFREMER ftp
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
spec2D <- get_2Dspectrum("SEMREVO", start = "1994-01-01", end = "1994-02-28")
image(spec2D$dir, spec2D$freq, spec2D$efth[, , 1],
  xlab = "Direction (°)",
  ylab = "Frequency (Hz"
)
\dontshow{\}) # examplesIf}
}
