% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_parameters_raw}
\alias{get_parameters_raw}
\title{Get time series of sea-state parameters}
\usage{
get_parameters_raw(
  parameter = "hs",
  node = 42,
  start = as.POSIXct("1994-01-01Z00:00:00"),
  end = as.POSIXct("1994-12-31Z23:00:00")
)
}
\arguments{
\item{parameter}{variable to fetch from the database}

\item{node}{single integer with the node to get}

\item{start}{starting date (posixct)}

\item{end}{end date (posixct)}
}
\value{
a tibble with 2 columns and as many rows as needed
}
\description{
No verification is made within this function, and thus it should not be called directly by the user
}
