% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Value Matching}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{value to search}

\item{table}{table of values}
}
\value{
the opposite of x \%in\% table
}
\description{
Value Matching
}
\examples{
1:10 \%in\% c(1, 3, 5, 9)
1:10 \%nin\% c(1, 3, 5, 9)
}
