% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specta_data_ploting.R
\name{plot_1d_specta}
\alias{plot_1d_specta}
\title{Plot a wave density 1D spectrum at a given time}
\usage{
plot_1d_specta(spec, time = 1L, print_sea_state = TRUE, ...)
}
\arguments{
\item{spec}{the spectral data, as an output from \code{get_2Dspectrum}}

\item{time}{the time to plot. Either an integer or the date.}

\item{print_sea_state}{should the sea_states parameters being plot ? Default to TRUE.}

\item{...}{currently unused}
}
\value{
a ggplot object
}
\description{
Plot a wave density 1D spectrum at a given time
}
\examples{
plot_1d_specta(resourcecodedata::rscd_1d_spectra, 1)
}
