% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_coastline}
\alias{rscd_coastline}
\title{Resourcecode coastline}
\format{
\subsection{\code{rscd_coastline}}{

A data frame with 24403 rows and 3 columns:
\describe{
\item{longitude,latitude}{coordinates of the border line}
\item{depth}{depth of the border.}
}
}
}
\source{
User Manual of the RESOURCECODE database \url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_coastline
}
\description{
This data contains the coastline used to run the RESOURCECODE hindcast. This will be mainly used for ploting purpose.
}
\keyword{datasets}
