% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_data_example}
\alias{rscd_data_example}
\title{Resourcecode data extract}
\format{
\subsection{A data frame with 87,648 rows and 7 variables:}{

\describe{
\item{time}{POSIXct. Timestamp in UTC corresponding to the observation time (hourly resolution).}
\item{hs}{numeric. Significant wave height (m).}
\item{tp}{numeric. Peak wave period (s).}
\item{dp}{numeric. Mean wave direction (degrees, coming from true North, clockwise).}
\item{uwnd}{numeric. Zonal (east–west) component of the 10-m wind (m/s). Positive eastward.}
\item{vwnd}{numeric. Meridional (north–south) component of the 10-m wind (m/s). Positive northward.}
\item{dpt}{numeric.Depth (m).}
}
}
}
\source{
User Manual of the RESOURCECODE database
\url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_data_example
}
\description{
An extract of the Resourcecode Hindcast database, at node 123456, spanning from
1994-01-01 00:00:00 to 1999-12-31 23:00:00.
}
\keyword{datasets}
