% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_dir}
\alias{rscd_dir}
\title{Resourcecode directional bins}
\format{
\subsection{\code{rscd_dir}}{

An array of length 36 with the directionnal bins
}
}
\source{
User Manual of the RESOURCECODE database
\url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_dir
}
\description{
(equivalent to a directional resolution of 10°;
}
\keyword{datasets}
