% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_field}
\alias{rscd_field}
\title{Resourcecode FIELD grid}
\format{
\subsection{\code{rscd_field}}{

A data frame with 328,030 rows and 5 columns:
\describe{
\item{node}{node number}
\item{longitude, latitude}{coordinates of the nodes}
\item{depth}{depth of the node}
\item{d50}{Bottom sediment types}
}
}
}
\source{
User Manual of the RESOURCECODE database
\url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_field
}
\description{
This data contains the location and characteristics of the 328,030 nodes
where the RESOURCECODE hindcast data is available
}
\keyword{datasets}
