% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_freq}
\alias{rscd_freq}
\title{Resourcecode frequency vector of 1D and 2D spectra}
\format{
\subsection{\code{rscd_triangles}}{

An array 36 elements with the frequencies values
}
}
\source{
User Manual of the RESOURCECODE database
\url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_freq
}
\description{
The wave spectrum discretization considers 36 frequencies, starting from
0.0339 Hz up to 0.9526 Hz with a frequency increment factor of 1.1
}
\keyword{datasets}
