% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_islands}
\alias{rscd_islands}
\title{Resourcecode islands coastline}
\format{
\subsection{\code{rscd_islands}}{

A data frame with 24403 rows and 3 columns:
\describe{
\item{longitude,latitude}{coordinates of the border line}
\item{depth}{depth of the border}
\item{ID}{Unique number used to identify the island}
}
}
}
\source{
User Manual of the RESOURCECODE database \url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_islands
}
\description{
This data contains the coastline of the islands used to run the RESOURCECODE hindcast, as data separated from the mainland.
This will be mainly used for ploting purpose.
}
\keyword{datasets}
