% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rscd_mapplot}
\alias{rscd_mapplot}
\title{Create a map of the provided variable on the RESOURCECODE field grid}
\usage{
rscd_mapplot(z, name = "Depth (m)", zlim = NULL)
}
\arguments{
\item{z}{the data ro plot: a vector of the same size as the grid (328,030 rows).}

\item{name}{name of the variable plored, to be included in the legend.}

\item{zlim}{limits of the scale. See \code{\link[ggplot2]{continuous_scale}} for details.}
}
\value{
a ggplot2 object
}
\description{
Create a map of the provided variable on the RESOURCECODE field grid
}
\examples{
\dontrun{
 rscd_mapplot(rscd_field$depth)
}
}
