% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_spectral}
\alias{rscd_spectral}
\title{Resourcecode SPEC grid}
\format{
\subsection{\code{rscd_spectral}}{

A data frame with 24,276 rows and 5 columns:
\describe{
\item{longitude, latitude}{coordinates of the nodes}
\item{name}{Name of the spectral output point}
\item{depth}{depth of the node}
\item{d50}{Bottom sediment types}
}
}
}
\source{
User Manual of the RESOURCECODE database
\url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_spectral
}
\description{
This data contains the location and characteristics of the 24,276 nodes
where the full 2D spectral data is available in the RESOURCECODE data.
}
\keyword{datasets}
