% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_triangles}
\alias{rscd_triangles}
\title{Resourcecode triangles}
\format{
\subsection{\code{rscd_triangles}}{

A matrix with 3 rows and 566506 columns:
\describe{
\item{rows}{verticies of the triangles}
\item{columns}{node number of each vertices}
}
}
}
\source{
User Manual of the RESOURCECODE database \url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_triangles
}
\description{
This data contains the triangles of the unstructured computational mesh.
This will be mainly used for ploting purpose.
}
\keyword{datasets}
