% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rscd_variables}
\alias{rscd_variables}
\title{Resourcecode variable list}
\format{
\subsection{\code{rscd_variables}}{

A data frame with 88 rows and 3 columns:
\describe{
\item{name}{short name of the variable}
\item{longname}{Full name}
\item{unit}{Unit}
}
}
}
\source{
User Manual of the RESOURCECODE database \url{https://archimer.ifremer.fr/doc/00751/86306/}
}
\usage{
rscd_variables
}
\description{
This data contains the variables available in the FIELD database.
}
\keyword{datasets}
