% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_data_download.R
\name{get_1Dspectrum}
\alias{get_1Dspectrum}
\title{Download the 1D spectrum data from IFREMER ftp}
\usage{
get_1Dspectrum(point, start = "1994-01-01", end = "1994-02-28")
}
\arguments{
\item{point}{the location name (string) requested. The consistency is checked internally.}

\item{start}{the starting date (as a string). The consistency is checked internally.}

\item{end}{the ending date as a string}
}
\value{
A list with the data read from the downloaded netCDF file.
}
\description{
Download the 1D spectrum data from IFREMER ftp
}
\examples{
spec = get_1Dspectrum("SEMREVO",start="1994-01-01",end="1994-02-28")
}
