% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{jonswap}
\alias{jonswap}
\title{Compute Jonswap spectrum with f (Hz) formulation (Sf = 2\emph{pi}Sw)}
\usage{
jonswap(hs, tp, gamma = 3.3, freq)
}
\arguments{
\item{hs}{Significant wave height (m)}

\item{tp}{Peak period (s)}

\item{gamma}{the peakness factor (e.g. 1 or 3.3)}

\item{freq}{the frequency vector where the spectrum is to be computed (Hz)}
}
\value{
a vector containing the spectrum on input freq
}
\description{
Compute Jonswap spectrum with f (Hz) formulation (Sf = 2\emph{pi}Sw)
}
\examples{
Sp = jonswap(1,10,3.3)
}
