% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcecode-package.R
\docType{package}
\name{resourcecode-package}
\alias{resourcecode}
\alias{resourcecode-package}
\title{resourcecode: Access to the 'RESOURCECODE' Hindcast Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to download data from the 'RESOURCECODE' hindcast database of sea-states, time series of sea-state parameters and time series of 1D and 2D wave spectra. See \url{https://resourcecode.ifremer.fr} for more details about the available data. Also provides facilities to plot and analyse downloaded data, such as computing the sea-state parameters from both the 1D and 2D surface elevation variance spectral density.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Resourcecode-project/r-resourcecode}
  \item \url{https://resourcecode-project.github.io/r-resourcecode/}
  \item \url{https://resourcecode-project.r-universe.dev/resourcecode}
  \item Report bugs at \url{https://github.com/Resourcecode-project/r-resourcecode/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Raillard \email{nicolas.raillard@ifremer.fr} (\href{https://orcid.org/0000-0003-3385-5104}{ORCID})

Authors:
\itemize{
  \item Nicolas Raillard \email{nicolas.raillard@ifremer.fr} (\href{https://orcid.org/0000-0003-3385-5104}{ORCID})
}

}
\keyword{internal}
