% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcecode-package.R
\docType{package}
\name{resourcecode-package}
\alias{resourcecode}
\alias{resourcecode-package}
\title{resourcecode: Access to the RESOURCECODE hindcast database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

WThis package provides utility functions to download data from the RESOURCECODE database, along with facilities to plot and analyse such data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NRaillard/resourcecode}
  \item \url{https://nraillard.github.io/resourcecode/}
  \item Report bugs at \url{https://github.com/NRaillard/resourcecode/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Raillard \email{nicolas.raillard@ifremer.fr} (\href{https://orcid.org/0000-0003-3385-5104}{ORCID})

}
\keyword{internal}
